+++ subsystems/investment/__init__.py
@@ -1,0 +1,3 @@
+from subsystems.investment.subsystem import InvestmentSubsystem
+
+__all__ = ["InvestmentSubsystem"]